return minimal unsigned index types
The SAIS algorithm uses negative values but once the suffix array
is computed, we can save space by using unsigned types for indices
and shrinking the type down to the smallest unsigned type we can.
Depending on the length of the array, we may be able to do this by
reinterpreting the indices or we may have to copy them.